Suppress a warning in a case where we throw away a mutex guard.
(Clippy is complaining because there is no point in locking
something then throwing away the guard; it immediately releases the
lock, which is probably not what you wanted.)

Remember,
```
  let _ = (...);
```
and
```
  let _ignore = (...);
```
have different semantics:  The former drops the result of the expression
immediately, whereas the latter drops it when it goes out of scope.

This code still isn't right, but clippy isn't complaining any more.
The right fix is to use crossbeam; see #27.